LayoutTests/inspector/runtime/executionContextCreated-onEnable-expected.txt


Test that exactly one Runtime.executionContextCreated event is fired for each existing context when Runtime.enable is called.


== Running test suite: Runtime.executionContextCreated.OnEnable
-- Running test case: Runtime.executionContextCreated.OnEnable.NoDuplicates
Execution context created: id=0 frameId=0 isPageContext=true
Execution context created: id=1 frameId=1 isPageContext=true
PASS: Should receive 2 executionContextCreated events (for main world in the main frame and the subframe).
PASS: Should have 2 unique contexts.
